requirements.txt
torch
torchvision
flask
opencv-python
numpy
pillow